修改类型

+++ com-center-service/src/main/java/com/center/service/controller/UserController.java
@@ -27,7 +27,7 @@ public class UserController {
 
     @GetMapping("/detail/get")
     @ApiOperation(value = "获取用户详情接口")
-    public UserDto detailGet(@RequestParam(name = "id")Integer id){
+    public UserDto detailGet(@RequestParam(name = "id")Long id){
         return userService.selectByPrimaryKey(id);
     }
